resolve dependency cycle error
