+++ src/components/Form/FormTabSelect/index.tsx
@@ -1,5 +1,4 @@
 /* eslint-disable @typescript-eslint/no-use-before-define */
-/* eslint-disable import/no-cycle */
 import React, { useEffect } from 'react';
 import styled from 'styled-components';
